Updated R code for 9.2 and 9.4
So when there's only code with contextless inputs, we leave it in the corresponding section; when there are practice exercises, we put them in a new section called "Practice Exercises". Sometimes a question can have multiple parts that combine different sections, and putting similar questions together helps students compare and contrast. Would this arrangement be a bit confusing though?